login, logout and registration page

+++ base/views.py
@@ -1,14 +1,77 @@
 from django.shortcuts import render, redirect
+from django.contrib import messages
 from django.http import HttpResponse
-from . models import Room
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from django.contrib.auth import login, logout, authenticate
+from django.db.models import Q
+from . models import Room, Topic
 from . forms import RoomForm
+from django.contrib.auth.forms import UserCreationForm
 
 # Create your views here.
 
+def loginPage(request):
+    
+    page = "login"
+    
+    if request.user.is_authenticated:
+        return redirect("home")
+    
+    
+    if request.method == "POST":
+        username = request.POST.get("username").lower()
+        password = request.POST.get("password")
+        
+        user = authenticate(request, username=username, password=password)
+        
+        if user is not None:
+            login(request, user)
+            return redirect("home")
+        else:
+            messages.error(request, "Username or Password is not correct")
+            
+    context ={"page":page}        
+    return render(request, "base/login_register.html", context)
+            
+            
+def logoutUser(request):
+    logout(request)
+    return redirect("home") 
+            
+def registerPage(request):
+    form = UserCreationForm()
+    
+    if request.method =="POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            return redirect("login")
+        else:
+            messages.error(request, "An error occured during registration")
+    
+    
+    context ={"form":form}        
+    return render(request, "base/login_register.html", context)
 
 def home(request):
-    rooms = Room.objects.all()
-    context = {"rooms":rooms,}
+    q = request.GET.get("q") if request.GET.get("q") != None else ""
+    rooms = Room.objects.filter(
+        Q(topic__name__icontains=q) |
+        Q(name__icontains=q) |
+        Q(host__username__icontains=q)
+        )
+    topics = Topic.objects.all()
+    count_room = rooms.count()
+    
+    
+    context = {
+        "rooms":rooms,
+        "topics": topics,
+        "count_room": count_room
+        }
     return render(request, "base/home.html", context)
 
 
@@ -17,6 +80,7 @@ def room(request, pk):
     context = {"room":room}
     return render(request, "base/room.html", context)
 
+@login_required(login_url="login")
 def createRoom(request):
     form = RoomForm()
     
@@ -29,9 +93,15 @@ def createRoom(request):
     context ={"form":form}
     return render(request, "base/room_form.html", context)
 
+@login_required(login_url="login")
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user.username != room.host.username:
+        return HttpResponse("You are not permitted here")
+        
+        
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -40,8 +110,14 @@ def updateRoom(request, pk):
     context  = {"form": form}
     return render(request, "base/room_form.html", context)
 
+@login_required(login_url="login")
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user.username != room.host.username:
+        print("request:", request.user)
+        print("user:", room.host.username)
+        return HttpResponse("You are not permitted here")
    
     if request.method == "POST":
         room.delete()
